+++ CHANGES.md
@@ -20,6 +20,7 @@ obedience.
 * Fixed error in TownMeet kidnap scenario
 * Image reset if no file could be loaded
 * Correctly present stat/skill changes after loading girl
+* Fixed some errors in lua scripts
  
 # Version 7.1.2beta3
 ## Gameplay